Refactor `MergeFunction` package and update imports.
Moved `MergeFunction` to a new package `caching.merge` to improve module structure and organization. Updated all relevant imports in `CacheContext` and `CacheFactory` to reflect the change.
