+++ assembler/src/main/java/io/github/pellse/assembler/caching/factory/CacheContext.java
@@ -18,6 +18,7 @@ package io.github.pellse.assembler.caching.factory;
 
 import io.github.pellse.assembler.RuleMapperContext.OneToManyContext;
 import io.github.pellse.assembler.RuleMapperContext.OneToOneContext;
+import io.github.pellse.assembler.caching.merge.MergeFunction;
 import io.github.pellse.assembler.caching.factory.CacheFactory.CacheTransformer;
 
 import java.util.*;
